feat(types): began migrating to new type module format
rather than a monolithic `Material` class we now have a wrapper class providing generic methods which is subclassed by specific materials with concrete implementations. this makes testing a less daunting prospect. currently we have implemented the logic for the local problem, and are in train on the nonlocal version.

BREAKING CHANGE: